added usable ProgressBar to Map and Statistic

+++ app/src/main/java/de/dhbw/corona_world_app/ui/tools/LoadingScreenInterface.java
@@ -10,5 +10,5 @@ public interface LoadingScreenInterface {
 
     public void setProgressBar(int progress, @NonNull String message);
 
-    public int getProgress();
+    public float getProgress();
 }
